refactor: split models.py to note.py and category.py
